backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
python-docx
pypdf2
openpyxl
pandas
Pillow
pytesseract
pdf2image
psycopg[binary]
pgvector
